Move get_user_appointments and set_target_user logic to helper method
