+++ app/models/user.rb
@@ -9,22 +9,6 @@ class User < ApplicationRecord
   has_many :appointments, dependent: :destroy
   has_many :inverse_appointments, class_name: 'Appointment', foreign_key: 'doctor_id'
 
-  # def get_appointment(id, url)
-  #   if patient_url(url)
-  #     appointments.find(id)
-  #   elsif doctor_url(url)
-  #     inverse_appointments.find(id)
-  #   end
-  # end
-
-  def get_appointments(_id, url)
-    if patient_url(url)
-      appointments.includes(:doctor).to_json({ include: :doctor })
-    elsif doctor_url(url)
-      inverse_appointments.includes(:user).to_json({ include: :user })
-    end
-  end
-
   def doctor_url(url)
     regex = %r{/api/v1/doctors}
     regex.match(url)
